(no message)
SCCS-vsn: usr.bin/tip/tip.h 5.8

+++ usr/src/usr.bin/tip/tip.h
@@ -1,10 +1,11 @@
 /*
- * Copyright (c) 1983 The Regents of the University of California.
+ * Copyright (c) 1989 The Regents of the University of California.
  * All rights reserved.
+ *
  *
  * %sccs.include.redist.c%
  *
- *	@(#)tip.h	5.7 (Berkeley) %G%
+ *      @(#)tip.h	5.8 (Berkeley) %G%
  */
 
 /*
@@ -45,8 +46,8 @@ char	*PH;			/* phone number file */
 char	*RM;			/* remote file name */
 char	*HO;			/* host name */
 
-int	BR;			/* line speed for conversation */
-int	FS;			/* frame size for transfers */
+long	BR;			/* line speed for conversation */
+long	FS;			/* frame size for transfers */
 
 char	DU;			/* this host is dialed up */
 char	HW;			/* this device is hardwired, see hunt.c */
@@ -56,9 +57,9 @@ char	*FO;			/* force (literal next) char*/
 char	*RC;			/* raise character */
 char	*RE;			/* script record file */
 char	*PR;			/* remote prompt */
-int	DL;			/* line delay for file transfers to remote */
-int	CL;			/* char delay for file transfers to remote */
-int	ET;			/* echocheck timeout */
+long	DL;			/* line delay for file transfers to remote */
+long	CL;			/* char delay for file transfers to remote */
+long	ET;			/* echocheck timeout */
 char	HD;			/* this host is half duplex - do local echo */
 
 /*
